Some testing added.
