+++ src/triggered/OpenCompas.ts
@@ -1,6 +1,6 @@
 import {LitElement} from 'lit-element';
 import {newWizardEvent} from '../foundation.js';
-import {compasSclTypeListWizard} from "../compas/CompasScltype.js";
+import {compasSclTypeListWizard} from "../compas/CompasScltypeList.js";
 
 export default class OpenCompasPlugin extends LitElement {
   async trigger(): Promise<void> {
